+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [BREAKING] Replaced the `frontend-url` param for `frontend-bind-port` ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
 - [BREAKING] Added `no-frontend` param to optionally disable the frontend server ([#156](https://github.com/0xMiden/miden-faucet/pull/156)).
 - Redesigned the frontend ([#201](https://github.com/0xMiden/miden-faucet/pull/201)).
+- Added `note-transport-url` CLI param to set the note transport layer ([#191](https://github.com/0xMiden/miden-faucet/pull/191)).
 
 ## 0.12.4 (2025-12-04)